requirements.txt
torchvision==0.18.0
torchaudio==2.3.0
hydra-core==1.3.0
flwr[simulation]>=1.0, <2.0
numpy==1.26.0
pandas==2.2.0
